fix maths
