+++ packages/repocop/src/metrics.test.ts
@@ -1,0 +1,18 @@
+import { getPercentageTrue } from './metrics';
+
+describe('When evaluating the percentage of trues, an empty array', () => {
+	test('should return a value of zero', () => {
+		expect(getPercentageTrue([])).toBe(0);
+	});
+});
+describe('When evaluating the proportion of trues and falses, we', () => {
+	test('should return a value between zero and 100', () => {
+		expect(getPercentageTrue([true])).toBe(100);
+		expect(getPercentageTrue([false])).toBe(0);
+		expect(getPercentageTrue([true, false])).toBe(50);
+	});
+	test('Should round to the nearest integer value', () => {
+		expect(getPercentageTrue([true, true, false])).toBe(67);
+		expect(getPercentageTrue([true, false, false])).toBe(33);
+	});
+});
